+++ contentful_management/asset.py
@@ -1,4 +1,4 @@
-from .resource import FieldsResource, PublishResource, ArchiveResource
+from .resource import FieldsResource, PublishResource, ArchiveResource, EnvironmentAwareResource
 
 
 """
@@ -14,7 +14,7 @@ API reference: https://www.contentful.com/developers/docs/references/content-man
 """
 
 
-class Asset(FieldsResource, PublishResource, ArchiveResource):
+class Asset(FieldsResource, PublishResource, ArchiveResource, EnvironmentAwareResource):
     """
     API reference: https://www.contentful.com/developers/docs/references/content-management-api/#/reference/assets
     """
@@ -49,7 +49,11 @@ class Asset(FieldsResource, PublishResource, ArchiveResource):
         for locale in self._fields.keys():
             self._client._put(
                 "{0}/files/{1}/process".format(
-                    self.__class__.base_url(self.space.id, self.id),
+                    self.__class__.base_url(
+                        self.space.id,
+                        self.id,
+                        environment_id=self._environment_id
+                    ),
                     locale
                 ),
                 {},