+++ contentful_management/environment_content_types_proxy.py
@@ -1,12 +1,12 @@
-from .space_resource_proxy import SpaceResourceProxy
+from .environment_resource_proxy import EnvironmentResourceProxy
 from .content_types_proxy import ContentTypesProxy
 
 
 """
-contentful_management.space_content_types_proxy
+contentful_management.environment_content_types_proxy
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-This module implements the SpaceContentTypesProxy class.
+This module implements the EnvironmentContentTypesProxy class.
 
 API reference: https://www.contentful.com/developers/docs/references/content-management-api/#/reference/content-types
 
@@ -15,7 +15,7 @@ API reference: https://www.contentful.com/developers/docs/references/content-man
 """
 
 
-class SpaceContentTypesProxy(SpaceResourceProxy):
+class EnvironmentContentTypesProxy(EnvironmentResourceProxy):
     """
     API reference: https://www.contentful.com/developers/docs/references/content-management-api/#/reference/content-types
     """