+++ tests/content_type_resource_proxy_test.py
@@ -6,4 +6,4 @@ from .test_helper import CLIENT, PLAYGROUND_SPACE
 class ContentTypeResourceProxyTest(TestCase):
     def test_resource_proxy_must_be_implemented(self):
         with self.assertRaises(Exception):
-            ContentTypeResourceProxy(CLIENT, PLAYGROUND_SPACE, 'foo')
+            ContentTypeResourceProxy(CLIENT, PLAYGROUND_SPACE, 'master', 'foo')