+++ tests/entries_proxy_test.py
@@ -7,7 +7,7 @@ from .test_helper import CLIENT, PLAYGROUND_SPACE
 class EntriesProxyTest(TestCase):
     @vcr.use_cassette('fixtures/entry/create_with_ct_empty.yaml')
     def test_entries_proxy_create_with_content_type_and_empty_attributes(self):
-        proxy = EntriesProxy(CLIENT, PLAYGROUND_SPACE, 'foo')
+        proxy = EntriesProxy(CLIENT, PLAYGROUND_SPACE, 'master', 'foo')
 
         entry = proxy.create('id_create_content_type_empty')