+++ tests/resource_test.py
@@ -15,9 +15,9 @@ class LinkTest(TestCase):
 
     @vcr.use_cassette('fixtures/link/resolve.yaml')
     def test_link_resolve(self):
-        content_type = CLIENT.content_types(PLAYGROUND_SPACE).find('foo')
+        content_type = CLIENT.content_types(PLAYGROUND_SPACE, 'master').find('foo')
 
-        resolved_ct = content_type.to_link().resolve(PLAYGROUND_SPACE)
+        resolved_ct = content_type.to_link().resolve(PLAYGROUND_SPACE, 'master')
 
         self.assertEqual(resolved_ct.id, 'foo')
         self.assertEqual(str(resolved_ct), str(content_type))